Refactor EditStudentCommand.java
