+++ src/main/java/seedu/address/logic/parser/EditStudentCommandParser.java
@@ -19,8 +19,8 @@ import java.util.Set;
 
 import seedu.address.commons.core.index.Index;
 import seedu.address.logic.commands.EditCommand;
-import seedu.address.logic.commands.EditStudentCommand;
 import seedu.address.logic.commands.descriptors.EditStudentDescriptor;
+import seedu.address.logic.commands.student.EditStudentCommand;
 import seedu.address.logic.parser.exceptions.ParseException;
 import seedu.address.model.tag.Tag;
